requirements-dev.txt
black==25.1.0
pytest==8.4.0
model-bakery==1.20.5
pytest-django==4.11.1
faker==37.4.0
faker-commerce==1.0.4
